+++ fireREST/fmc/object/certificatemap/__init__.py
@@ -1,0 +1,6 @@
+from fireREST.fmc import Resource
+
+
+class CertificateMap(Resource):
+    PATH = '/object/certificatemaps/{uuid}'
+    MINIMUM_VERSION_REQUIRED_GET = '7.0.0'