+++ fireREST/fmc/policy/identitypolicy/__init__.py
@@ -1,0 +1,6 @@
+from fireREST.fmc import Resource
+
+
+class IdentityPolicy(Resource):
+    PATH = '/policy/identitypolicies/{uuid}'
+    MINIMUM_VERSION_REQUIRED_GET = '7.0.0'